Renamed crowdsource-lab to radiolab-read-credit and the previous version to radiolab-read-credit. Revisions also include a modal window to enable readers to rehearse the script.
